main: move parsing of flags to init()
To avoid problems like parsing them multiple times.
